Move dtypes to separate module
