+++ edugrad/_tensor/tensor_reduce.py
@@ -2,7 +2,9 @@
 
 from __future__ import annotations
 
-from edugrad.helpers import dtypes, prod, all_int
+
+from edugrad.dtypes import dtypes
+from edugrad.helpers import prod, all_int
 from edugrad.function import Function
 import edugrad.function as function
